+++ src/main/java/com/invadermonky/futurefireproof/FutureFireproof.java
@@ -1,8 +1,10 @@
 package com.invadermonky.futurefireproof;
 
+import com.invadermonky.futurefireproof.config.ModTags;
 import com.invadermonky.futurefireproof.registry.RegistrarFF;
 import com.invadermonky.futurefireproof.util.LogHelper;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 
 @Mod(
@@ -14,7 +16,7 @@ import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 public class FutureFireproof {
     public static final String MOD_ID = "futurefireproof";
     public static final String MOD_NAME = "Future Fireproof";
-    public static final String VERSION = "1.12.2-1.1.0";
+    public static final String VERSION = "1.12.2-1.2.0";
     public static final String MC_VERSION = "[1.12.2]";
 
     @Mod.Instance(MOD_ID)
@@ -25,5 +27,11 @@ public class FutureFireproof {
         LogHelper.info("Starting " + MOD_NAME);
         RegistrarFF.registerEntities();
         LogHelper.debug("Finished preInit phase.");
+    }
+
+    @Mod.EventHandler
+    public void postInit(FMLPostInitializationEvent event) {
+        ModTags.syncConfig();
+        LogHelper.debug("Finished postInit phase.");
     }
 }