+++ src/main/java/com/invadermonky/futurefireproof/config/ModTags.java
@@ -1,33 +1,36 @@
 package com.invadermonky.futurefireproof.config;
 
 import gnu.trove.map.hash.THashMap;
+import gnu.trove.set.hash.THashSet;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.DamageSource;
 import net.minecraft.util.math.MathHelper;
 
+import java.util.Arrays;
+
 public class ModTags {
-    public static THashMap<String, Integer> FIREPROOF_ITEMS;
+    public static THashSet<String> DAMAGE_TYPES = new THashSet<>();
+    public static THashMap<String, Integer> FIREPROOF_ITEMS = new THashMap<>();
+
+    public static boolean isIgnoredDamageType(DamageSource source) {
+        return DAMAGE_TYPES.contains(source.getDamageType()) && source != DamageSource.OUT_OF_WORLD;
+    }
 
     public static boolean isFireproofItem(ItemStack stack) {
-        if(FIREPROOF_ITEMS == null) {
-            syncConfig();
-        }
         String itemId = stack.getItem().getRegistryName().toString();
         return FIREPROOF_ITEMS.containsKey(itemId) || FIREPROOF_ITEMS.containsKey(itemId + ":" + stack.getMetadata());
     }
 
     public static int getFireproofDecayRate(ItemStack stack) {
-        if(FIREPROOF_ITEMS == null) {
-            syncConfig();
-        }
         String itemId = stack.getItem().getRegistryName().toString();
         return Math.max(FIREPROOF_ITEMS.getOrDefault(itemId, 1), FIREPROOF_ITEMS.getOrDefault(itemId + ":" + stack.getMetadata(), 1));
     }
 
     public static void syncConfig() {
-        if(FIREPROOF_ITEMS == null) {
-            FIREPROOF_ITEMS = new THashMap<>(ConfigHandlerFF.fireproofItems.length);
-        }
+        DAMAGE_TYPES.clear();
         FIREPROOF_ITEMS.clear();
+
+        DAMAGE_TYPES.addAll(Arrays.asList(ConfigHandlerFF.damageTypes));
 
         for(String str : ConfigHandlerFF.fireproofItems) {
             if(str.matches(".+=[0-9]+")) {