+++ changelog.md
@@ -1,5 +1,11 @@
 # Future Fireproof Changelog
 
+## 1.12.2-1.2.0
+- Added configuration option to make fireproof items immune to specific types of damage
+- Added configuration option to make fireproof items immune to explosions
+- Fixed a bug sometimes causing Fireproof items to be completely immune to damage
+  <br><br>
+
 ## 1.12.2-1.1.0
 - Added Realistic Item Drops compatibility
 - Items that extend `IFireproofItem` with a decay rate of 1 can no longer be enchanted with the Fireproof enchant